Fix compilation
