+++ web/apps/payments/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.js";
-import "./index.css";
+import App from "./App";
+import "./styles/globals.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
     <React.StrictMode>
